// src/constants/enums.ts
export enum BankMaps {
  ALL = 'mapbox://styles/iandyone/cllw9d9pg00gf01r49ilh84cz',
  NONE = 'mapbox://styles/iandyone/cllwb75il00gq01r49glw5j40',
  PRIOR = 'mapbox://styles/iandyone/cllw8h6uk00ha01pf2q327g48',
  ALPHA = 'mapbox://styles/iandyone/cllw9lf8l00gn01pjbzcub87a',
}

export enum CurrenciesList {
  EUR = 'EUR',
  USD = 'USD',
  BYN = 'BYN',
  PLN = 'PLN',
  GBP = 'GBP',
  CNY = 'CNY',
  CHF = 'CHF',
  AUD = 'AUD',
  UAH = 'UAH',
  JPY = 'JPY',
}

export enum Routes {
  MAIN = '/',
  TIMELINE = '/timeline',
  BANKS = '/banks',
  CONTACTS = '/contacts',
  UNKNOWN = '*',
}